Put top-level classes inside Git::Pr::Release module
