+++ lib/git/pr/release.rb
@@ -1,4 +1,6 @@
 require "git/pr/release/util"
+require "git/pr/release/pull_request"
+require "git/pr/release/dummy_pull_request"
 require "git/pr/release/cli"
 
 module Git
